requirements.txt
absl-py==1.2.0
appnope==0.1.3
asttokens==2.0.7
astunparse==1.6.3
azure-core==1.25.0
azure-functions==1.11.2
azure-identity==1.10.0
azure-storage-blob==12.13.1
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
deprecation==2.1.0
entrypoints==0.4
executing==0.9.1
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gensim==3.8.3
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
isodate==0.6.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==5.4.2
Pillow==9.2.0
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==3.14.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF2==2.10.0
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
smart-open==6.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
